+++ src/CollectdWinService/Properties/AssemblyInfo.cs
@@ -35,4 +35,4 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: AssemblyVersion("0.10.5.*")]
+[assembly: AssemblyVersion("0.10.6.*")]